feat(compiler): working state of extracted compiler
