+++ benchmarks/latest/rollup.config.mjs
@@ -22,6 +22,7 @@ export default [{
           'metadata',
           'expression-parser',
           'runtime',
+          'template-compiler',
           'runtime-html',
           'platform',
           'platform-browser',
@@ -48,6 +49,7 @@ export default [{
           'metadata',
           'expression-parser',
           'runtime',
+          'template-compiler',
           'runtime-html',
           'platform',
           'platform-browser',
