+++ packages-tooling/vite-plugin/src/index.ts
@@ -49,6 +49,7 @@ export default function au(options: {
         'dialog',
         'expression-parser',
         'runtime',
+        'template-compiler',
         'runtime-html',
         'router-lite',
       ].reduce((aliases, pkg) => {
