+++ packages/__e2e__/router/webpack.config.js
@@ -33,6 +33,7 @@ module.exports = function () {
           'kernel',
           'expression-parser',
           'runtime',
+          'template-compiler',
           'runtime-html',
         ].reduce((map, pkg) => {
           const name = `@aurelia/${pkg}`;
