+++ packages/__e2e__/select-safari16/vite.config.mjs
@@ -15,6 +15,7 @@ export default defineConfig({
         'kernel',
         'expression-parser',
         'runtime',
+        'template-compiler',
         'runtime-html',
         'router',
       ].reduce((map, pkg) => {
