+++ packages/runtime-html/src/renderer.ts
@@ -10,16 +10,13 @@ import {
 import {
   IExpressionParser,
   type ExpressionType,
-  type Interpolation,
-  type IsBindingBehavior,
   type AnyBindingExpression,
-  type ForOfStatement,
 } from '@aurelia/expression-parser';
 import {
   IObserverLocator,
   type IObservable,
 } from '@aurelia/runtime';
-import { toView, type BindingMode } from './binding/interfaces-bindings';
+import { BindingMode, defaultMode, toView } from './binding/interfaces-bindings';
 import { AttributeBinding } from './binding/attribute';
 import { InterpolationBinding } from './binding/interpolation-binding';
 import { ContentBinding } from "./binding/content-binding";
@@ -34,346 +31,21 @@ import { Controller, ICustomElementController, ICustomElementViewModel, IControl
 import { IPlatform } from './platform';
 import { IViewFactory } from './templating/view';
 import { IRendering } from './templating/rendering';
-import type { AttrSyntax } from './resources/attribute-pattern';
-import { objectKeys, isString, etIsProperty, etInterpolation, etIsIterator, etIsFunction, objectFreeze } from './utilities';
+import { objectKeys, isString, etIsProperty, etInterpolation, etIsIterator, etIsFunction } from './utilities';
 import { createInterface, registerResolver, singletonRegistration } from './utilities-di';
-import { IAuSlotProjections, IAuSlotsInfo, AuSlotsInfo } from './templating/controller.projection';
+import { IAuSlotsInfo, AuSlotsInfo } from './templating/controller.projection';
 
 import type { IHydratableController } from './templating/controller';
-import type { PartialCustomElementDefinition } from './resources/custom-element';
 import { ErrorNames, createMappedError } from './errors';
 import { SpreadBinding } from './binding/spread-binding';
+import { AttributeBindingInstruction, HydrateAttributeInstruction, HydrateElementInstruction, HydrateLetElementInstruction, HydrateTemplateController, IInstruction, ITemplateCompiler, InstructionType, InterpolationInstruction, IteratorBindingInstruction, LetBindingInstruction, ListenerBindingInstruction, PropertyBindingInstruction, RefBindingInstruction, SetAttributeInstruction, SetClassAttributeInstruction, SetPropertyInstruction, SetStyleAttributeInstruction, SpreadBindingInstruction, StylePropertyBindingInstruction, TextBindingInstruction } from '@aurelia/template-compiler';
 
-/** @internal */ export const hydrateElement = 'ra';
-/** @internal */ export const hydrateAttribute = 'rb';
-/** @internal */ export const hydrateTemplateController = 'rc';
-/** @internal */ export const hydrateLetElement = 'rd';
-/** @internal */ export const setProperty = 're';
-/** @internal */ export const interpolation = 'rf';
-/** @internal */ export const propertyBinding = 'rg';
-/** @internal */ export const letBinding = 'ri';
-/** @internal */ export const refBinding = 'rj';
-/** @internal */ export const iteratorBinding = 'rk';
-/** @internal */ export const multiAttr = 'rl';
-/** @internal */ export const textBinding = 'ha';
-/** @internal */ export const listenerBinding = 'hb';
-/** @internal */ export const attributeBinding = 'hc';
-/** @internal */ export const stylePropertyBinding = 'hd';
-/** @internal */ export const setAttribute = 'he';
-/** @internal */ export const setClassAttribute = 'hf';
-/** @internal */ export const setStyleAttribute = 'hg';
-/** @internal */ export const spreadBinding = 'hs';
-/** @internal */ export const spreadElementProp = 'hp';
-
-export const InstructionType = /*@__PURE__*/ objectFreeze({
-  hydrateElement,
-  hydrateAttribute,
-  hydrateTemplateController,
-  hydrateLetElement,
-  setProperty,
-  interpolation,
-  propertyBinding,
-  letBinding,
-  refBinding,
-  iteratorBinding,
-  multiAttr,
-  textBinding,
-  listenerBinding,
-  attributeBinding,
-  stylePropertyBinding,
-  setAttribute,
-  setClassAttribute,
-  setStyleAttribute,
-  spreadBinding,
-  spreadElementProp,
-});
-export type InstructionType = typeof InstructionType[keyof typeof InstructionType];
-
-export interface IInstruction {
-  readonly type: string;
-}
-export const IInstruction = /*@__PURE__*/createInterface<IInstruction>('Instruction');
-
-export function isInstruction(value: unknown): value is IInstruction {
-  const type = (value as { type?: string }).type;
-  return isString(type) && type.length === 2;
-}
-
-export class InterpolationInstruction {
-  public readonly type = interpolation;
-
-  public constructor(
-    public from: string | Interpolation,
-    public to: string,
-  ) {}
-}
-
-export class PropertyBindingInstruction {
-  public readonly type = propertyBinding;
-
-  public constructor(
-    public from: string | IsBindingBehavior,
-    public to: string,
-    public mode: BindingMode,
-  ) {}
-}
-
-export class IteratorBindingInstruction {
-  public readonly type = iteratorBinding;
-
-  public constructor(
-    public forOf: string | ForOfStatement,
-    public to: string,
-    public props: MultiAttrInstruction[],
-  ) {}
-}
-
-export class RefBindingInstruction {
-  public readonly type = refBinding;
-
-  public constructor(
-    public readonly from: string | IsBindingBehavior,
-    public readonly to: string
-  ) {}
-}
-
-export class SetPropertyInstruction {
-  public readonly type = setProperty;
-
-  public constructor(
-    public value: unknown,
-    public to: string,
-  ) {}
-}
-
-export class MultiAttrInstruction {
-  public readonly type = multiAttr;
-
-  public constructor(
-    public value: string,
-    public to: string,
-    public command: string | null,
-  ) {}
-}
-
-export class HydrateElementInstruction<T extends Record<PropertyKey, unknown> = Record<PropertyKey, unknown>> {
-  public readonly type = hydrateElement;
-
-  public constructor(
-    /**
-     * The name of the custom element this instruction is associated with
-     */
-    // in theory, Constructor of resources should be accepted too
-    // though it would be unnecessary right now
-    public res: string | /* Constructable |  */CustomElementDefinition,
-    /**
-     * Bindable instructions for the custom element instance
-     */
-    public props: IInstruction[],
-    /**
-     * Indicates what projections are associated with the element usage
-     */
-    public projections: Record<string, PartialCustomElementDefinition> | null,
-    /**
-     * Indicates whether the usage of the custom element was with a containerless attribute or not
-     */
-    public containerless: boolean,
-    /**
-     * A list of captured attr syntaxes
-     */
-    public captures: AttrSyntax[] | undefined,
-    /**
-     * Any data associated with this instruction
-     */
-    public readonly data: T,
-  ) {
-  }
-}
-
-export class HydrateAttributeInstruction {
-  public readonly type = hydrateAttribute;
-
-  public constructor(
-    // in theory, Constructor of resources should be accepted too
-    // though it would be unnecessary right now
-    public res: string | /* Constructable |  */CustomAttributeDefinition,
-    public alias: string | undefined,
-    /**
-     * Bindable instructions for the custom attribute instance
-     */
-    public props: IInstruction[],
-  ) {}
-}
-
-export class HydrateTemplateController {
-  public readonly type = hydrateTemplateController;
-
-  public constructor(
-    public def: PartialCustomElementDefinition,
-    // in theory, Constructor of resources should be accepted too
-    // though it would be unnecessary right now
-    public res: string | /* Constructable |  */CustomAttributeDefinition,
-    public alias: string | undefined,
-    /**
-     * Bindable instructions for the template controller instance
-     */
-    public props: IInstruction[],
-  ) {}
-}
-
-export class HydrateLetElementInstruction {
-  public readonly type = hydrateLetElement;
-
-  public constructor(
-    public instructions: LetBindingInstruction[],
-    public toBindingContext: boolean,
-  ) {}
-}
-
-export class LetBindingInstruction {
-  public readonly type = letBinding;
-
-  public constructor(
-    public from: string | IsBindingBehavior | Interpolation,
-    public to: string,
-  ) {}
-}
-
-export class TextBindingInstruction {
-  public readonly type = textBinding;
-
-  public constructor(
-    public from: string | IsBindingBehavior,
-  ) {}
-}
-
-export class ListenerBindingInstruction {
-  public readonly type = listenerBinding;
-
-  public constructor(
-    public from: string | IsBindingBehavior,
-    public to: string,
-    public capture: boolean,
-    public modifier: string | null,
-  ) {}
-}
-export class StylePropertyBindingInstruction {
-  public readonly type = stylePropertyBinding;
-
-  public constructor(
-    public from: string | IsBindingBehavior,
-    public to: string,
-  ) {}
-}
-
-export class SetAttributeInstruction {
-  public readonly type = setAttribute;
-
-  public constructor(
-    public value: string,
-    public to: string,
-  ) {}
-}
-
-export class SetClassAttributeInstruction {
-  public readonly type: typeof InstructionType.setClassAttribute = setClassAttribute;
-
-  public constructor(
-    public readonly value: string,
-  ) {}
-}
-
-export class SetStyleAttributeInstruction {
-  public readonly type: typeof InstructionType.setStyleAttribute = setStyleAttribute;
-
-  public constructor(
-    public readonly value: string,
-  ) {}
-}
-
-export class AttributeBindingInstruction {
-  public readonly type = attributeBinding;
-
-  public constructor(
-    /**
-     * `attr` and `to` have the same value on a normal attribute
-     * Will be different on `class` and `style`
-     * on `class`: attr = `class` (from binding command), to = attribute name
-     * on `style`: attr = `style` (from binding command), to = attribute name
-     */
-    public attr: string,
-    public from: string | IsBindingBehavior,
-    public to: string,
-  ) {}
-}
-
-export class SpreadBindingInstruction {
-  public readonly type = spreadBinding;
-}
-
-export class SpreadElementPropBindingInstruction {
-  public readonly type = spreadElementProp;
-  public constructor(
-    public readonly instructions: IInstruction,
-  ) {}
-}
-
-export const ITemplateCompiler = /*@__PURE__*/createInterface<ITemplateCompiler>('ITemplateCompiler');
-export interface ITemplateCompiler {
-  /**
-   * Indicates whether this compiler should compile template in debug mode
-   *
-   * For the default compiler, this means all expressions are kept as is on the template
-   */
-  debug: boolean;
-  /**
-   * Experimental API, for optimization.
-   *
-   * `true` to create CustomElement/CustomAttribute instructions
-   * with resolved resources constructor during compilation, instead of name
-   */
-  resolveResources: boolean;
-  compile(
-    partialDefinition: PartialCustomElementDefinition,
-    context: IContainer,
-    compilationInstruction: ICompliationInstruction | null,
-  ): PartialCustomElementDefinition;
-
-  /**
-   * Compile a list of captured attributes as if they are declared in a template
-   *
-   * @param requestor - the context definition where the attributes is compiled
-   * @param attrSyntaxes - the attributes captured
-   * @param container - the container containing information for the compilation
-   * @param host - the host element where the attributes are spreaded on
-   */
-  compileSpread(
-    requestor: CustomElementDefinition,
-    attrSyntaxes: AttrSyntax[],
-    container: IContainer,
-    target: Element,
-    /**
-     * An associated custom element definition for the target host element
-     * Sometimes spread compilation may occur without the container having all necessary information
-     * about the targeted element that is receiving the spread
-     *
-     * Caller of this method may want to provide this information dynamically instead
-     */
-    targetDef?: CustomElementDefinition,
-  ): IInstruction[];
-}
-
-export interface ICompliationInstruction {
-  /**
-   * A record of projections available for compiling a template.
-   * Where each key is the matching slot name for <au-slot/> inside,
-   * and each value is the definition to render and project
-   */
-  projections: IAuSlotProjections | null;
-}
-
-export interface IRenderer<TType extends string = string> {
-  target: TType;
+/**
+ * An interface describing an instruction renderer
+ * its target property will be used to match instruction types dynamically at render time
+ */
+export interface IRenderer {
+  target: string;
   render(
     /**
      * The controller that is current invoking this renderer
@@ -389,7 +61,7 @@ export interface IRenderer<TType extends string = string> {
 
 export const IRenderer = /*@__PURE__*/createInterface<IRenderer>('IRenderer');
 
-export function renderer<TType extends string, T extends IRenderer<TType>, C extends Constructable<T>>(target: C, context: ClassDecoratorContext): C {
+export function renderer<T extends IRenderer, C extends Constructable<T>>(target: C, context: ClassDecoratorContext): C {
   return Registrable.define(target, function (this: typeof target, container: IContainer): void {
     singletonRegistration(IRenderer, this).register(container);
   });
@@ -436,8 +108,8 @@ function getRefTarget(refHost: INode, refTargetName: string): object {
   }
 }
 
-export const SetPropertyRenderer = /*@__PURE__*/ renderer(class SetPropertyRenderer implements IRenderer<'re'> {
-  public readonly target = setProperty;
+export const SetPropertyRenderer = /*@__PURE__*/ renderer(class SetPropertyRenderer implements IRenderer {
+  public readonly target = InstructionType.setProperty;
 
   public render(
     renderingCtrl: IHydratableController,
@@ -456,12 +128,12 @@ export const SetPropertyRenderer = /*@__PURE__*/ renderer(class SetPropertyRende
 export const CustomElementRenderer = /*@__PURE__*/ renderer(class CustomElementRenderer implements IRenderer {
   /** @internal */ public readonly _rendering = resolve(IRendering);
 
-  public readonly target = hydrateElement;
+  public readonly target = InstructionType.hydrateElement;
 
   public render(
     renderingCtrl: IHydratableController,
     target: HTMLElement,
-    instruction: HydrateElementInstruction,
+    instruction: HydrateElementInstruction<Record<PropertyKey, unknown>, CustomElementDefinition>,
     platform: IPlatform,
     exprParser: IExpressionParser,
     observerLocator: IObserverLocator,
@@ -531,7 +203,7 @@ export const CustomElementRenderer = /*@__PURE__*/ renderer(class CustomElementR
 export const CustomAttributeRenderer = /*@__PURE__*/ renderer(class CustomAttributeRenderer implements IRenderer {
   /** @internal */ public readonly _rendering = resolve(IRendering);
 
-  public readonly target = hydrateAttribute;
+  public readonly target = InstructionType.hydrateAttribute;
 
   public render(
     /**
@@ -562,7 +234,7 @@ export const CustomAttributeRenderer = /*@__PURE__*/ renderer(class CustomAttrib
       //   def = CustomAttribute.getDefinition(instruction.res);
       //   break;
       default:
-        def = instruction.res;
+        def = instruction.res as CustomAttributeDefinition;
     }
     const results = invokeAttribute(
       /* platform         */platform,
@@ -601,7 +273,7 @@ export const CustomAttributeRenderer = /*@__PURE__*/ renderer(class CustomAttrib
 export const TemplateControllerRenderer = /*@__PURE__*/ renderer(class TemplateControllerRenderer implements IRenderer {
   /** @internal */ public readonly _rendering = resolve(IRendering);
 
-  public readonly target = hydrateTemplateController;
+  public readonly target = InstructionType.hydrateTemplateController;
 
   public render(
     renderingCtrl: IHydratableController,
@@ -629,7 +301,7 @@ export const TemplateControllerRenderer = /*@__PURE__*/ renderer(class TemplateC
       //   def = CustomAttribute.getDefinition(instruction.res);
       //   break;
       default:
-        def = instruction.res;
+        def = instruction.res as CustomAttributeDefinition;
     }
     // const viewFactory = this._rendering.getViewFactory(
     //   instruction.def,
@@ -679,7 +351,7 @@ export const TemplateControllerRenderer = /*@__PURE__*/ renderer(class TemplateC
 }, null!);
 
 export const LetElementRenderer = /*@__PURE__*/ renderer(class LetElementRenderer implements IRenderer {
-  public readonly target = hydrateLetElement;
+  public readonly target = InstructionType.hydrateLetElement;
   public render(
     renderingCtrl: IHydratableController,
     target: Node & ChildNode,
@@ -713,7 +385,7 @@ export const LetElementRenderer = /*@__PURE__*/ renderer(class LetElementRendere
 }, null!);
 
 export const RefBindingRenderer = /*@__PURE__*/ renderer(class RefBindingRenderer implements IRenderer {
-  public readonly target = refBinding;
+  public readonly target = InstructionType.refBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: INode,
@@ -730,7 +402,7 @@ export const RefBindingRenderer = /*@__PURE__*/ renderer(class RefBindingRendere
 }, null!);
 
 export const InterpolationBindingRenderer = /*@__PURE__*/ renderer(class InterpolationBindingRenderer implements IRenderer {
-  public readonly target = interpolation;
+  public readonly target = InstructionType.interpolation;
   public render(
     renderingCtrl: IHydratableController,
     target: IController,
@@ -753,7 +425,7 @@ export const InterpolationBindingRenderer = /*@__PURE__*/ renderer(class Interpo
 }, null!);
 
 export const PropertyBindingRenderer = /*@__PURE__*/ renderer(class PropertyBindingRenderer implements IRenderer {
-  public readonly target = propertyBinding;
+  public readonly target = InstructionType.propertyBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: IController,
@@ -762,6 +434,7 @@ export const PropertyBindingRenderer = /*@__PURE__*/ renderer(class PropertyBind
     exprParser: IExpressionParser,
     observerLocator: IObserverLocator,
   ): void {
+    const mode = isString(instruction.mode) ? BindingMode[instruction.mode as keyof typeof BindingMode] ?? defaultMode : instruction.mode as BindingMode;
     renderingCtrl.addBinding(new PropertyBinding(
       renderingCtrl,
       renderingCtrl.container,
@@ -770,13 +443,13 @@ export const PropertyBindingRenderer = /*@__PURE__*/ renderer(class PropertyBind
       ensureExpression(exprParser, instruction.from, etIsProperty),
       getTarget(target),
       instruction.to,
-      instruction.mode,
+      mode,
     ));
   }
 }, null!);
 
 export const IteratorBindingRenderer = /*@__PURE__*/ renderer(class IteratorBindingRenderer implements IRenderer {
-  public readonly target = iteratorBinding;
+  public readonly target = InstructionType.iteratorBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: IController,
@@ -799,7 +472,7 @@ export const IteratorBindingRenderer = /*@__PURE__*/ renderer(class IteratorBind
 }, null!);
 
 export const TextBindingRenderer = /*@__PURE__*/ renderer(class TextBindingRenderer implements IRenderer {
-  public readonly target = textBinding;
+  public readonly target = InstructionType.textBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: ChildNode,
@@ -834,7 +507,7 @@ export const IListenerBindingOptions = createInterface<IListenerBindingOptions>(
 }));
 
 export const ListenerBindingRenderer = /*@__PURE__*/ renderer(class ListenerBindingRenderer implements IRenderer {
-  public readonly target = listenerBinding;
+  public readonly target = InstructionType.listenerBinding;
 
   /** @internal */
   public readonly _modifierHandler = resolve(IEventModifier);
@@ -860,7 +533,7 @@ export const ListenerBindingRenderer = /*@__PURE__*/ renderer(class ListenerBind
 }, null!);
 
 export const SetAttributeRenderer = /*@__PURE__*/ renderer(class SetAttributeRenderer implements IRenderer {
-  public readonly target = setAttribute;
+  public readonly target = InstructionType.setAttribute;
   public render(
     _: IHydratableController,
     target: HTMLElement,
@@ -871,7 +544,7 @@ export const SetAttributeRenderer = /*@__PURE__*/ renderer(class SetAttributeRen
 }, null!);
 
 export const SetClassAttributeRenderer = /*@__PURE__*/ renderer(class SetClassAttributeRenderer implements IRenderer {
-  public readonly target = setClassAttribute;
+  public readonly target = InstructionType.setClassAttribute;
   public render(
     _: IHydratableController,
     target: HTMLElement,
@@ -882,7 +555,7 @@ export const SetClassAttributeRenderer = /*@__PURE__*/ renderer(class SetClassAt
 }, null!);
 
 export const SetStyleAttributeRenderer = /*@__PURE__*/ renderer(class SetStyleAttributeRenderer implements IRenderer {
-  public readonly target = setStyleAttribute;
+  public readonly target = InstructionType.setStyleAttribute;
   public render(
     _: IHydratableController,
     target: HTMLElement,
@@ -918,7 +591,7 @@ const ambiguousStyles = [
 ];
 
 export const StylePropertyBindingRenderer = /*@__PURE__*/ renderer(class StylePropertyBindingRenderer implements IRenderer {
-  public readonly target = stylePropertyBinding;
+  public readonly target = InstructionType.stylePropertyBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: HTMLElement,
@@ -968,7 +641,7 @@ class DevStylePropertyBinding extends PropertyBinding {
 }
 
 export const AttributeBindingRenderer = /*@__PURE__*/ renderer(class AttributeBindingRenderer implements IRenderer {
-  public readonly target = attributeBinding;
+  public readonly target = InstructionType.attributeBinding;
   public render(
     renderingCtrl: IHydratableController,
     target: HTMLElement,
@@ -1002,7 +675,7 @@ export const SpreadRenderer = /*@__PURE__*/ renderer(class SpreadRenderer implem
   /** @internal */ public readonly _compiler = resolve(ITemplateCompiler);
   /** @internal */ public readonly _rendering = resolve(IRendering);
 
-  public readonly target = spreadBinding;
+  public readonly target = InstructionType.spreadBinding;
 
   public render(
     renderingCtrl: IHydratableController,
