+++ packages/runtime-html/src/resources/custom-elements/au-compose.ts
@@ -3,7 +3,6 @@ import { IObserverLocator } from '@aurelia/runtime';
 import { Scope } from '../../binding/scope';
 import { INode, IRenderLocation, convertToRenderLocation, registerHostNode } from '../../dom';
 import { IPlatform } from '../../platform';
-import { HydrateElementInstruction, IInstruction, ITemplateCompiler } from '../../renderer';
 import { Controller, HydrationContext, IController, ICustomElementController, IHydratedController, IHydrationContext, ISyntheticView, vmkCe } from '../../templating/controller';
 import { IRendering } from '../../templating/rendering';
 import { isFunction, isPromise } from '../../utilities';
@@ -12,7 +11,7 @@ import { CustomElement, CustomElementDefinition, CustomElementStaticAuDefinition
 import { ErrorNames, createMappedError } from '../../errors';
 import { fromView } from '../../binding/interfaces-bindings';
 import { SpreadBinding } from '../../binding/spread-binding';
-import { AttrSyntax } from '../attribute-pattern';
+import { HydrateElementInstruction, IInstruction, ITemplateCompiler, AttrSyntax } from '@aurelia/template-compiler';
 import { IExpressionParser } from '@aurelia/expression-parser';
 
 /**
