+++ packages/runtime-html/src/resources/template-controllers/switch.ts
@@ -18,7 +18,7 @@ import { isArray } from '../../utilities';
 
 import type { Controller, ICustomAttributeController, ICustomAttributeViewModel, IHydratedController, IHydratedParentController, IHydratableController, ISyntheticView, ControllerVisitor } from '../../templating/controller';
 import type { INode } from '../../dom';
-import type { IInstruction } from '../../renderer';
+import type { IInstruction } from '@aurelia/template-compiler';
 import { createMappedError, ErrorNames } from '../../errors';
 import { PartialBindableDefinition } from '../../bindable';
