+++ packages/template-compiler/.eslintignore
@@ -1,0 +1,4 @@
+dist
+node_modules
+.eslintrc.cjs
+*.js
