+++ packages/template-compiler/.eslintrc.cjs
@@ -1,0 +1,15 @@
+const path = require('path');
+const thisDir = path.resolve(__dirname);
+
+module.exports = {
+  extends: [
+    '../../.eslintrc.cjs',
+  ],
+  parserOptions: {
+    project: path.join(thisDir, 'tsconfig.json'),
+    tsconfigRootDir: thisDir,
+  },
+  rules: {
+    '@typescript-eslint/no-explicit-any': 'off',
+  }
+};
