+++ packages/template-compiler/README.md
@@ -1,0 +1,19 @@
+[![License: MIT](https://img.shields.io/badge/License-MIT-yellow.svg)](https://opensource.org/licenses/MIT)
+[![TypeScript](https://img.shields.io/badge/%3C%2F%3E-TypeScript-%230074c1.svg)](http://www.typescriptlang.org/)
+[![CircleCI](https://circleci.com/gh/aurelia/aurelia.svg?style=shield)](https://circleci.com/gh/aurelia/aurelia)
+[![npm](https://img.shields.io/npm/v/@aurelia/metadata.svg?maxAge=3600)](https://www.npmjs.com/package/@aurelia/metadata)
+# @aurelia/template-compiler
+
+## Installing
+
+For the latest stable version:
+
+```bash
+npm i @aurelia/template-compiler
+```
+
+For our nightly builds:
+
+```bash
+npm i @aurelia/template-compiler@dev
+```
