+++ packages/template-compiler/rollup.config.mjs
@@ -1,0 +1,5 @@
+// @ts-check
+import pkg from './package.json' assert { type: 'json' };
+import { getRollupConfig } from '../rollup-utils.mjs';
+
+export default getRollupConfig(pkg);
