+++ packages/template-compiler/src/index.ts
@@ -1,0 +1,102 @@
+export {
+  type IAttributeComponentDefinition,
+  type ICompiledElementComponentDefinition,
+  type IComponentBindablePropDefinition,
+  type IElementComponentDefinition,
+  type ProcessContentHook,
+  type IDomPlatform,
+  ITemplateCompiler,
+} from './interfaces-template-compiler';
+
+export {
+  IAttrMapper,
+  type IsTwoWayPredicate
+} from './attribute-mapper';
+
+export {
+  type AttributePatternDefinition,
+  type AttributePatternKind,
+  IAttributePattern,
+  IAttributeParser,
+  AttributeParser,
+  ISyntaxInterpreter,
+  SyntaxInterpreter,
+  AttrSyntax,
+  AttributePattern,
+  Interpretation,
+  ColonPrefixedBindAttributePattern,
+  AtPrefixedTriggerAttributePattern,
+  DotSeparatedAttributePattern,
+  EventAttributePattern,
+  RefAttributePattern,
+  SpreadAttributePattern,
+  attributePattern,
+} from './attribute-pattern';
+
+export {
+  type PartialBindingCommandDefinition,
+  type BindingCommandDecorator,
+  BindingCommandDefinition,
+  type BindingCommandInstance,
+  type BindingCommandStaticAuDefinition,
+  type BindingCommandKind,
+  type IBindableCommandInfo,
+  type ICommandBuildInfo,
+  type IPlainAttrCommandInfo,
+  BindingCommand,
+  AttrBindingCommand,
+  BindingCommandType,
+  CaptureBindingCommand,
+  ClassBindingCommand,
+  DefaultBindingCommand,
+  ForBindingCommand,
+  FromViewBindingCommand,
+  OneTimeBindingCommand,
+  RefBindingCommand,
+  SpreadBindingCommand,
+  StyleBindingCommand,
+  ToViewBindingCommand,
+  TriggerBindingCommand,
+  TwoWayBindingCommand,
+  bindingCommand,
+} from './binding-command';
+
+export {
+  IResourceResolver,
+  IBindablesInfoResolver,
+  type IElementBindablesInfo,
+  type IAttributeBindablesInfo,
+  ITemplateCompilerHooks,
+  TemplateCompiler,
+  TemplateCompilerHooks,
+  templateCompilerHooks,
+} from './template-compiler';
+
+export {
+  ITemplateElementFactory,
+} from './template-element-factory';
+
+export {
+  AttributeBindingInstruction,
+  HydrateAttributeInstruction,
+  HydrateElementInstruction,
+  HydrateLetElementInstruction,
+  HydrateTemplateController,
+  IInstruction,
+  InstructionType,
+  InterpolationInstruction,
+  IteratorBindingInstruction,
+  LetBindingInstruction,
+  ListenerBindingInstruction,
+  MultiAttrInstruction,
+  PropertyBindingInstruction,
+  RefBindingInstruction,
+  SetAttributeInstruction,
+  SetClassAttributeInstruction,
+  SetPropertyInstruction,
+  SetStyleAttributeInstruction,
+  SpreadBindingInstruction,
+  SpreadElementPropBindingInstruction,
+  StylePropertyBindingInstruction,
+  TextBindingInstruction
+} from './instructions';
