+++ packages/template-compiler/tsconfig.build.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../tsconfig-build.json",
+  "compilerOptions": {
+    "composite": false,
+    "declaration": false,
+    "declarationMap": false,
+    "rootDir": "src",
+    "lib": ["esnext", "DOM"]
+  },
+  "include": [
+    "src"
+  ]
+}
