+++ packages/template-compiler/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../tsconfig-build.json",
+  "compilerOptions": {
+    "declarationDir": "dist/types",
+    "outDir": "dist/esm",
+    "rootDir": "src",
+    "lib": ["esnext", "dom"]
+  },
+  "include": ["src"]
+}
